+++ NEWS.md
@@ -1,3 +1,9 @@
+# bcgovr 1.0.5
+
+* Added function `insert_bcgov_lifecycle_badge()` to access new [bcgov project lifecycle badges](https://github.com/bcgov/repomountie/blob/master/doc/lifecycle-badges.md) 
+* Deprecated `insert_bcgov_devex_badge()` and removed reference to `devex_badge()`
+* Updated **Insert bcgov Lifecycle Badge** RStudio addin 
+
 # bcgovr 1.0.4
 
 * fixed several bugs introduced by changes in `usethis`; predominately around project switching/activation.